Reorganise project into folders
