+++ adapters/orm.py
@@ -7,10 +7,9 @@ the other way around).
 See https://docs.sqlalchemy.org/en/20/orm/mapping_styles.html#imperative-mapping for docs on this
 manual mapping style.
 """
+from domain import model
 from sqlalchemy import Column, Date, Float, Integer, String, Table
 from sqlalchemy.orm import registry
-
-import model
 
 mapper_registry = registry()
