+++ setup.cfg
@@ -35,7 +35,7 @@ warn_no_return = true
 show_error_codes = true
 
 # Don't require test functions to include types.
-[mypy-test_model,test_orm,test_interface,test_repository,test_service,test_usecases,conftest]
+[mypy-tests.*]
 allow_untyped_defs = true
 disable_error_code = attr-defined
